Refactor database module boundaries
